Beginning of support for data symbols
This is somewhat limited because local symbols may be a substantial portion
of the data section.
Also, currently bloaty includes space taken by a symbol's name in the
name section in its attribution for the symbol (which can be larger
than its space in the data section). This seems WAI for bloaty
but might not be what we want.